Rename ConsensusService to ConsensusController
